# packages/contracts/openapi/aurim-first-slice.v0.yaml
openapi: 3.1.0
info:
  title: Aurim First Slice API
  version: 0.1.0
  description: >
    Contract-first API for the initial Aurim vertical slice covering workspaces,
    git workspace bindings, factory specs, factory runs, and factory artifacts.
servers:
  - url: /
...
paths:
...
  /factory-runs/{factory_run_id}/artifacts:
    post:
      tags: [Factory Artifacts]
      operationId: createFactoryArtifact
      parameters:
        - $ref: "#/components/parameters/FactoryRunId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/create-factory-artifact-request.schema.json
      responses:
        "201":
          description: Factory artifact created.
          content:
            application/json:
              schema:
                $ref: ../schemas/factory-artifact.schema.json
    get:
      tags: [Factory Artifacts]
      operationId: listFactoryArtifacts
      parameters:
        - $ref: "#/components/parameters/FactoryRunId"
      responses:
        "200":
          description: Factory artifacts for the run.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: ../schemas/factory-artifact.schema.json
  /factory-artifacts/{factory_artifact_id}:
    get:
      tags: [Factory Artifacts]
      operationId: getFactoryArtifact
      parameters:
        - $ref: "#/components/parameters/FactoryArtifactId"
      responses:
        "200":
          description: Factory artifact details.
          content:
            application/json:
              schema:
                $ref: ../schemas/factory-artifact.schema.json
components:
  parameters:
...
    FactoryRunId:
      name: factory_run_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FactoryArtifactId:
      name: factory_artifact_id
      in: path
      required: true
      schema:
        type: string
        format: uuid